Moved module into separate file, corrected function calls
